pyproject.toml
[project]
name = "stream-score-stack"
version = "0.1.0"
description = "Streamlit Airflow Postres project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = []

[dependency-groups]
airflow = [
    "apache-airflow>=3.0.0",
    "apache-airflow-providers-postgres>=6.1.3",
]
core = [
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "sqlalchemy>=1.4.54",
    "streamlit>=1.45.0",
]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.8",
]
